LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/parsing/column-rule-computed-expected.txt


FAIL Property column-rule value '10px' assert_equals: expected "10px rgb(0, 255, 0)" but got "0px none rgb(0, 255, 0)"
PASS Property column-rule value 'dotted'
FAIL Property column-rule value '0px none rgb(255, 0, 255)' assert_equals: expected "0px rgb(255, 0, 255)" but got "0px none rgb(255, 0, 255)"
PASS Property column-rule value '10px dotted rgb(255, 0, 255)'
FAIL Property column-rule value 'medium hidden currentcolor' assert_equals: expected "3px hidden rgb(0, 255, 0)" but got "0px hidden rgb(0, 255, 0)"
PASS Property column-rule value 'medium solid currentcolor'
